Updated flux value number formatting
As an intermediate solution, I have set it up so that if a number is
less than 0.0001 or greater than 10000 (absolute value), the cell is
formatted using scientific notation. Also I set a minimum cell width of
90 which allows exactly 12 digits to display. The largest number that
can be entered without scientific notation is 9999.99999999 and it fits
exactly into the minimum width.

Also the scientific notation format is "#.######E0", so even
0.0000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000000111111111111
which is 1.111111E-125 fits into the cell with no clipping. So far I
haven't been able to break it.

The formula bar displays whatever is entered into the cell, although I
could convert it to always display a decimal format number if you want
that.